pyproject.toml
[tool.poetry]
name = "slack-monitor"
version = "1.0.0"
description = "Slack Monitor Service on Cognite Functions template for OIDC projects"
authors = ["ivan.surif@cognite.com"]

[tool.poetry.dependencies]
python = "^3.9"
pytest-custom-exit-code = ">=0.3"
cognite-sdk = ">=6"
azure-identity = "^1.14.1"
azure-storage-queue = "^12.7.3"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = ">=7"
pre-commit = ">=3"

[tool.isort]
profile = "black"
force_single_line = false
atomic = true
lines_after_imports = 2
lines_between_types = 1
line_length = 120
known_third_party = ["cognite","pytest"]

[tool.black]
line-length = 120
